Added default constructor to SignalHandle<T> with invalid state (nullptr) Created SignalSetBuilder with fluent API for resolving multiple signals Added Resolver::signals() factory method returning builder Created SubscriptionSetBuilder with fluent API for batch setup Added Client::subscriptions() factory method Changed subscribe() and serve_actuator() from Status to void   Only throws std::logic_error if client already started (programmer error)   Real errors reported at start() or wait_until_ready() (async registration)
Implemented transparent type mapping for int8/uint8/int16/uint16
Added widening (send) and narrowing (receive) conversions in vss_client.cpp
Added range validation for narrowing conversions

**Breaking Changes:**
  - subscribe() and serve_actuator() now return void instead of Status

+++ examples/climate_control/climate_control.hpp
@@ -92,8 +92,7 @@ private:
     kuksa::SignalHandle<float> cabin_temp_;                // Vehicle.Cabin.HVAC.Station.Row1.Driver.Temperature
 
     // VSS 5.1 Signal handles (outputs - protection actions)
-    // NOTE: VSS specifies uint8 for Window.Position, but KUKSA uses uint32 (see TYPE_SYSTEM_ANALYSIS.md)
-    kuksa::SignalHandle<uint32_t> window_position_;        // Vehicle.Cabin.Door.Row1.DriverSide.Window.Position
+    kuksa::SignalHandle<uint8_t> window_position_;         // Vehicle.Cabin.Door.Row1.DriverSide.Window.Position (VSS 5.1: uint8)
     kuksa::SignalHandle<std::string> sunroof_switch_;      // Vehicle.Cabin.Sunroof.Switch
 
     // Custom signals (vss_extensions.json)
